test/requirements.txt
Pillow
packaging
pyautogui
pystray
requests